+++ example/lib/domain/notifiers/example_filters/example_filters_provider.dart
@@ -3,15 +3,16 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 final exampleFiltersProvider = StateProvider<String?>((ref) => null);
 
 // This is also possible, subscription will work the same for StateProvider
-// and for StateNotifierProvider
+// and for NotifierProvider
 
 // final exampleFiltersProvider =
-//     StateNotifierProvider<ExampleFiltersStateNotifier, String?>(
-//   (ref) => ExampleFiltersStateNotifier(),
+//     NotifierProvider<ExampleFiltersNotifier, String?>(
+//   () => ExampleFiltersNotifier(),
 // );
 
-// class ExampleFiltersStateNotifier extends StateNotifier<String?> {
-//   ExampleFiltersStateNotifier() : super(null);
+// class ExampleFiltersNotifier extends SimpleNotifier<String?> {
+//   @override
+//   String? prepareForBuild() => null;
 
 //   void update(String? text) => state = text;
 // }
