+++ example/lib/domain/notifiers/example_notifier/example_state_notifier.dart
@@ -1,5 +1,6 @@
 // ignore_for_file: always_use_package_imports
 
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:q_architecture/base_state_notifier.dart';
 import 'package:q_architecture/q_architecture.dart';
 
@@ -7,14 +8,16 @@ import '../../../data/repositories/example_repository.dart';
 import '../example_filters/example_filters_provider.dart';
 
 final exampleNotifierProvider =
-    BaseStateNotifierProvider<ExampleStateNotifier, String>(
-  (ref) => ExampleStateNotifier(ref.watch(exampleRepositoryProvider), ref),
+    NotifierProvider<ExampleStateNotifier, BaseState<String>>(
+  () => ExampleStateNotifier(),
 );
 
-class ExampleStateNotifier extends BaseStateNotifier<String> {
-  final ExampleRepository _exampleRepository;
+class ExampleStateNotifier extends BaseNotifier<String> {
+  late ExampleRepository _exampleRepository;
 
-  ExampleStateNotifier(this._exampleRepository, super.ref) {
+  @override
+  void prepareForBuild() {
+    _exampleRepository = ref.watch(exampleRepositoryProvider);
     on(
       exampleFiltersProvider,
       (previous, next) => getSomeStringsStreamed(),
