+++ lib/src/domain/notifiers/simple_notifier.dart
@@ -1,137 +1,56 @@
 // ignore_for_file: always_use_package_imports
 
-import 'dart:async';
-
 import 'package:flutter/foundation.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:q_architecture/q_architecture.dart';
+import 'package:q_architecture/src/domain/mixins/simple_notifier_mixin.dart';
 
-mixin SimpleNotifierController {
-  late final Ref _ref;
-  Timer? _debounceTimer;
-  final Map<String, bool> _isThrottlingMap = {};
+abstract class SimpleNotifier<T> extends Notifier<T> with SimpleNotifierMixin {
+  T prepareForBuild();
 
-  void initWithRef(Ref ref) {
-    _ref = ref;
-    _ref.onDispose(() => _debounceTimer?.cancel());
+  /// do not override in child classes, use prepareForBuild instead
+  @nonVirtual
+  @override
+  T build() {
+    initWithRef(ref);
+    return prepareForBuild();
   }
-
-  ///Show [BaseLoadingIndicator] above the entire app
-  @protected
-  void showGlobalLoading() =>
-      _ref.read(globalLoadingProvider.notifier).update((state) => true);
-
-  ///Clear [BaseLoadingIndicator]
-  @protected
-  void clearGlobalLoading() =>
-      _ref.read(globalLoadingProvider.notifier).update((state) => false);
-
-  @protected
-  void setGlobalFailure(Failure? failure) {
-    clearGlobalLoading();
-    _ref
-        .read(globalFailureProvider.notifier)
-        .update((state) => failure?.copyWith(uniqueKey: UniqueKey()));
-  }
-
-  @protected
-  void setGlobalInfo(GlobalInfo? globalInfo) {
-    clearGlobalLoading();
-    _ref
-        .read(globalInfoProvider.notifier)
-        .update((state) => globalInfo?.copyWith(uniqueKey: UniqueKey()));
-  }
-
-  ///Subscribe to another notifier's state changes
-  @protected
-  void on<U>(
-    AlwaysAliveProviderListenable<U> provider,
-    void Function(U? previous, U next) invokeFunction, {
-    bool Function(U? previous, U next)? skipUpdateCondition,
-  }) =>
-      _ref.listen(
-        provider,
-        (previous, next) {
-          if (previous == next) return;
-          if (skipUpdateCondition?.call(previous, next) ?? false) return;
-
-          invokeFunction.call(previous, next);
-        },
-      );
-
-  ///Wait to collect multiple method calls for certain duration before allowing only one method call to proceed
-  @protected
-  Future<void> debounce({
-    Duration duration = const Duration(milliseconds: 500),
-  }) async {
-    if (_debounceTimer?.isActive == true) {
-      _debounceTimer?.cancel();
-    }
-    final debounceCompleter = Completer();
-    _debounceTimer = Timer(
-      duration,
-      () => debounceCompleter.complete(),
-    );
-    await debounceCompleter.future;
-  }
-
-  ///Execute given function and then block further executing of the same function for certain duration.
-  ///[waitForFunction] if set to true it will wait if function finishes after provided duration delay, otherwise will finish immediately after given duration
-  @protected
-  Future<void> throttle(
-    Future<void> Function() function, {
-    Duration duration = const Duration(milliseconds: 500),
-    bool waitForFunction = true,
-    String throttleIdentifier = '',
-  }) async {
-    if (_isThrottlingMap[throttleIdentifier] == true) return;
-    _isThrottlingMap[throttleIdentifier] = true;
-    final functionCompleter = Completer();
-    final throttleCompleter = Completer();
-    var durationFinished = false;
-    void completeThrottle() {
-      if (!throttleCompleter.isCompleted) {
-        _isThrottlingMap[throttleIdentifier] = false;
-        throttleCompleter.complete();
-      }
-    }
-
-    function().then(
-      (value) {
-        functionCompleter.complete();
-        if (durationFinished) completeThrottle();
-      },
-      onError: (error) {
-        functionCompleter.completeError(error);
-        if (durationFinished) completeThrottle();
-      },
-    );
-    Future.delayed(duration, () {
-      durationFinished = true;
-      if (!waitForFunction || functionCompleter.isCompleted) {
-        completeThrottle();
-      }
-    });
-    await Future.wait([
-      if (waitForFunction) functionCompleter.future,
-      throttleCompleter.future,
-    ]);
-  }
-
-  ///Cancels if throttling is in progress
-  @protected
-  void cancelThrottle({String throttleIdentifier = ''}) =>
-      _isThrottlingMap[throttleIdentifier] = false;
 }
 
-abstract class SimpleNotifier<T> extends Notifier<T>
-    with SimpleNotifierController {}
-
 abstract class AutoDisposeSimpleNotifier<T> extends AutoDisposeNotifier<T>
-    with SimpleNotifierController {}
+    with SimpleNotifierMixin {
+  T prepareForBuild();
 
-abstract class FamilySimpleNotifier<T, Param> extends FamilyNotifier<T, Param>
-    with SimpleNotifierController {}
+  /// do not override in child classes, use prepareForBuild instead
+  @nonVirtual
+  @override
+  T build() {
+    initWithRef(ref);
+    return prepareForBuild();
+  }
+}
 
-abstract class AutoDisposeFamilySimpleNotifier<T, Param>
-    extends AutoDisposeFamilyNotifier<T, Param> with SimpleNotifierController {}
+abstract class FamilySimpleNotifier<T, Arg> extends FamilyNotifier<T, Arg>
+    with SimpleNotifierMixin {
+  T prepareForBuild(Arg arg);
+
+  /// do not override in child classes, use prepareForBuild instead
+  @nonVirtual
+  @override
+  T build(Arg arg) {
+    initWithRef(ref);
+    return prepareForBuild(arg);
+  }
+}
+
+abstract class AutoDisposeFamilySimpleNotifier<T, Arg>
+    extends AutoDisposeFamilyNotifier<T, Arg> with SimpleNotifierMixin {
+  T prepareForBuild(Arg arg);
+
+  /// do not override in child classes, use prepareForBuild instead
+  @nonVirtual
+  @override
+  T build(Arg arg) {
+    initWithRef(ref);
+    return prepareForBuild(arg);
+  }
+}
