+++ lib/src/domain/notifiers/simple_state_notifier.dart
@@ -1,10 +1,10 @@
 // ignore_for_file: always_use_package_imports
 
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:q_architecture/q_architecture.dart';
+import 'package:q_architecture/src/domain/mixins/simple_notifier_mixin.dart';
 
 abstract class SimpleStateNotifier<T> extends StateNotifier<T>
-    with SimpleNotifierController {
+    with SimpleNotifierMixin {
   final Ref ref;
   SimpleStateNotifier(this.ref, T initialState) : super(initialState) {
     initWithRef(ref);
